Consolidate similar tests and setups
