--- EurofurenceTests/Application/Components/Event Detail/Presenter Tests/WhenBindingDealersDenBanner_EventDetailPresenterShould.swift
@@ -1,21 +1,0 @@
-import Eurofurence
-import EurofurenceModel
-import EurofurenceModelTestDoubles
-import XCTest
-
-class WhenBindingDealersDenBanner_EventDetailPresenterShould: XCTestCase {
-
-    func testBindTheMessageOntoTheComponent() {
-        let event = FakeEvent.random
-        let message = String.random
-        let artShowViewModel = EventDealersDenMessageViewModel(message: message)
-        let viewModel = StubDealersDenEventViewModel(dealersDenMessageViewModel: artShowViewModel)
-        let viewModelFactory = FakeEventDetailViewModelFactory(viewModel: viewModel, for: event)
-        let context = EventDetailPresenterTestBuilder().with(viewModelFactory).build(for: event)
-        context.simulateSceneDidLoad()
-        context.scene.bindComponent(at: IndexPath(item: 0, section: 0))
-
-        XCTAssertEqual(message, context.scene.stubbedDealersDenMessageComponent.capturedMessage)
-    }
-
-}
