--- EurofurenceTests/Application/Components/Event Detail/Presenter Tests/WhenBindingMainStageEvent_EventDetailPresenterShould.swift
@@ -1,21 +1,0 @@
-import Eurofurence
-import EurofurenceModel
-import EurofurenceModelTestDoubles
-import XCTest
-
-class WhenBindingMainStageEvent_EventDetailPresenterShould: XCTestCase {
-
-    func testBindTheMessageOntoTheComponent() {
-        let event = FakeEvent.random
-        let message = String.random
-        let kageMessageViewModel = EventMainStageMessageViewModel(message: message)
-        let viewModel = StubMainStageEventViewModel(mainStageMessageViewModel: kageMessageViewModel)
-        let viewModelFactory = FakeEventDetailViewModelFactory(viewModel: viewModel, for: event)
-        let context = EventDetailPresenterTestBuilder().with(viewModelFactory).build(for: event)
-        context.simulateSceneDidLoad()
-        context.scene.bindComponent(at: IndexPath(item: 0, section: 0))
-
-        XCTAssertEqual(message, context.scene.stubbedMainStageMessageComponent.capturedMessage)
-    }
-
-}
