--- EurofurenceTests/Application/Components/Event Detail/Presenter Tests/WhenBindingPhotoshootBanner_EventDetailPresenterShould.swift
@@ -1,21 +1,0 @@
-import Eurofurence
-import EurofurenceModel
-import EurofurenceModelTestDoubles
-import XCTest
-
-class WhenBindingPhotoshootBanner_EventDetailPresenterShould: XCTestCase {
-
-    func testBindTheMessageOntoTheComponent() {
-        let event = FakeEvent.random
-        let message = String.random
-        let kageMessageViewModel = EventPhotoshootMessageViewModel(message: message)
-        let viewModel = StubPhotoshootEventViewModel(photoshootMessageViewModel: kageMessageViewModel)
-        let viewModelFactory = FakeEventDetailViewModelFactory(viewModel: viewModel, for: event)
-        let context = EventDetailPresenterTestBuilder().with(viewModelFactory).build(for: event)
-        context.simulateSceneDidLoad()
-        context.scene.bindComponent(at: IndexPath(item: 0, section: 0))
-
-        XCTAssertEqual(message, context.scene.stubbedPhotoshootMessageComponent.capturedMessage)
-    }
-
-}
